requirements.txt
pandas==2.3.3
kaggle==1.7.4.5
python-dotenv==1.1.1
scikit-learn==1.7.2
dvc[s3]==3.63.0
pytest==8.4.2
mlflow==3.5.0
xgboost==3.1.0
torch==2.9.0
imblearn==0.0
pydantic==2.11.10
pwdlib[argon2]==0.3.0
uvicorn==0.38.0
slowapi==0.1.9
shap==0.49.1
evidently==0.7.15
prometheus_client==0.23.1
prometheus-fastapi-instrumentator==7.1.0